% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_builders.R
\name{build_combo}
\alias{build_combo}
\title{Build treatment-control combinations}
\usage{
build_combo(groups, method = "treatmentcontrol")
}
\arguments{
\item{groups}{A vector of unique treatment groups (assumed sorted and numeric)}

\item{method}{String: either "pairwise" or "treatmentcontrol"}
}
\value{
A matrix of treatment-control combinations
}
\description{
Build treatment-control combinations
}
