% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_builders.R
\name{build_D}
\alias{build_D}
\title{Build D matrix (treatment indicator)}
\usage{
build_D(data, treatment)
}
\arguments{
\item{data}{Data frame}

\item{treatment}{Character vector of outcome column names}
}
\value{
A matrix of outcomes
}
\description{
Build D matrix (treatment indicator)
}
