% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_builders.R
\name{build_DX}
\alias{build_DX}
\title{Build DX matrix (treatment + controls)}
\usage{
build_DX(data, treatment, controls = NULL)
}
\arguments{
\item{data}{Data frame}

\item{treatment}{Character vector of outcome column names}

\item{controls}{Optional character vector of control variable names}
}
\value{
A matrix of outcomes
}
\description{
Build DX matrix (treatment + controls)
}
