% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_output.R
\name{build_output}
\alias{build_output}
\title{Format final output from mhtexp2_r}
\usage{
build_output(observed, combo, alpha_sin, alpha_mul, pvals, alpha_mulm)
}
\arguments{
\item{observed}{3D array of observed stats (outcome x subgroup x comparison)}

\item{combo}{Matrix of treatment-control pairs}

\item{alpha_sin}{Matrix of single-hypothesis thresholds}

\item{alpha_mul}{Matrix of multiple-testing corrected thresholds}

\item{pvals}{Matrix of p-values}

\item{alpha_mulm}{Matrix of transitivity-corrected thresholds}
}
\value{
A long-format data.frame of test results
}
\description{
Format final output from mhtexp2_r
}
