% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_builders.R
\name{build_sub}
\alias{build_sub}
\title{Build subgroup vector}
\usage{
build_sub(data, subgroup, n = nrow(data))
}
\arguments{
\item{data}{Data frame}

\item{subgroup}{Optional character name of subgroup column}

\item{n}{Integer, number of rows}
}
\value{
Integer vector of subgroup IDs
}
\description{
Build subgroup vector
}
