% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_builders.R
\name{build_Y}
\alias{build_Y}
\title{Build Y matrix (outcomes)}
\usage{
build_Y(data, outcomes)
}
\arguments{
\item{data}{Data frame}

\item{outcomes}{Character vector of outcome column names}
}
\value{
A matrix of outcomes
}
\description{
Build Y matrix (outcomes)
}
