% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_control.R
\name{calculate_alphamul}
\alias{calculate_alphamul}
\title{Calculate multiple-testing adjusted thresholds (Theorem 3.1)
Calculate multiple-testing adjusted thresholds (Theorem 3.1) with subgroup support}
\usage{
calculate_alphamul(observed, boot)
}
\arguments{
\item{observed}{3D array of observed test stats (outcome x subgroup x comparison)}

\item{boot}{4D array of bootstrap stats (B x outcome x subgroup x comparison)}
}
\value{
Matrix of adjusted alpha thresholds (same shape as observed)
}
\description{
Calculate multiple-testing adjusted thresholds (Theorem 3.1)
Calculate multiple-testing adjusted thresholds (Theorem 3.1) with subgroup support
}
