% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtexp2_r.R
\name{mhtexp2_r}
\alias{mhtexp2_r}
\title{mhtexp2_r: R wrapper for multiple hypothesis testing procedure}
\usage{
mhtexp2_r(
  Y,
  treatment,
  controls = NULL,
  subgroup = NULL,
  combo = "treatmentcontrol",
  bootstrap = 3000,
  studentized = TRUE,
  transitivity_check = TRUE
)
}
\arguments{
\item{Y}{Matrix or data.frame of outcomes}

\item{treatment}{Vector of treatment assignments}

\item{controls}{Optional matrix or data.frame of controls}

\item{subgroup}{Optional vector of subgroup identifiers}

\item{combo}{"pairwise" or "treatmentcontrol"}

\item{bootstrap}{Number of bootstrap replications}

\item{studentized}{Logical, whether to studentize test stats (default = TRUE)}

\item{transitivity_check}{Logical, whether to apply transitivity correction (Remark 3.8)}
}
\value{
A list containing the output data.frame and intermediate arrays
}
\description{
mhtexp2_r: R wrapper for multiple hypothesis testing procedure
}
