% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_package.R
\name{run_stata_mhtexp2}
\alias{run_stata_mhtexp2}
\title{Run Stata mhtexp2 and return results}
\usage{
run_stata_mhtexp2(
  data,
  combo,
  bootstrap,
  studentized,
  transitivity_check,
  seed = NULL,
  stata_path,
  verbose = TRUE
)
}
\description{
Run Stata mhtexp2 and return results
}
