% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runreg.R
\name{runreg}
\alias{runreg}
\title{Run adjusted regression to estimate treatment effect}
\usage{
runreg(X, y, Xbar, Xvar, pi_2, pi_1, pi_z)
}
\arguments{
\item{X}{A matrix: first column is treatment dummy (1 for treatment, 0 for control), rest are controls (optional)}

\item{y}{A numeric vector of outcomes}

\item{Xbar}{A vector of covariate means for the subgroup (can be 0)}

\item{Xvar}{A covariance matrix of covariates (can be 0)}

\item{pi_2}{Proportion of treatment group}

\item{pi_1}{Proportion of control group}

\item{pi_z}{Proportion of the whole subgroup}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A named vector with ATE and SE
}\if{html}{\out{</div>}}
}
\description{
Run adjusted regression to estimate treatment effect
}
