% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivity_check.R
\name{violates_transitivity}
\alias{violates_transitivity}
\title{Check for transitivity violations in hypothesis set}
\usage{
violates_transitivity(subset_hypotheses, rejected)
}
\arguments{
\item{subset_hypotheses}{List of new hypotheses}

\item{rejected}{List of previously rejected hypotheses}
}
\value{
TRUE if contradiction found, FALSE otherwise
}
\description{
Check for transitivity violations in hypothesis set
}
