% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_helpers.R
\name{calculate_alphasin}
\alias{calculate_alphasin}
\title{Calculate single hypothesis adjusted thresholds (Remark 3.2)}
\usage{
calculate_alphasin(pact, pboot)
}
\arguments{
\item{pact}{3D array (outcomes x subgroups x comparisons) - observed "1-p" values}

\item{pboot}{4D array (B x outcomes x subgroups x comparisons) - bootstrap "1-p" values}
}
\value{
3D array of adjusted alpha thresholds (same shape)
}
\description{
Calculate single hypothesis adjusted thresholds (Remark 3.2)
}
